Update rack
addresses a security vulnerability notice on github

This repo is pretty stagnant and I would just "archive" it on github to
make the notice go away, but it is actually currently deployed to
heroku and I don't really want to take it down so I might as well patch
it.

This was a fun little project that I'll never finish.